feat (078): move AuthProperties into auth context
